requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
jinja2==3.1.2
starlette~=0.27.0
aiofiles~=24.1.0
pydantic~=1.10.22
pandas~=2.3.1